handled leading and interposed zero ansi-arguments
Since its valid ANSI to have a in other arguments embedded zero:
"\xB1[31;0;33m" we have to transform this into a end sequence and
a new sequence for restarting after linebreak.
All the while we have to ignore leading zeros, so that:
"\xB1[0100m" does not become "\xB1[0m\x1B100m".

After talking with muesli: Ignoring Write errors explicitly, maybe to be
changed in the future.